refactor: move set up spawner method to spawner manager class
