+++ Game/Spawner/SpawnManager.h
@@ -4,6 +4,16 @@
 #include "Game/Spawner/Spawner.h"
 #include "Game/Spawner/SpawnStrategy.h"
 #include "Game/Event.h"
+#include "Game/Spawner/SpawnComponent.h"
+
+struct EnemyParams {
+        int health;
+        float speed;
+        sf::Vector2f position;
+        std::string texturePath;
+        std::shared_ptr<AIControllerDifficulty> difficulty;
+        TimeBasedRule::TimeRule rule;
+};
 
 class SpawnManager
 {
@@ -13,5 +23,6 @@ class SpawnManager
         SpawnManager();
         ~SpawnManager() = default;
         void addStrategy(std::shared_ptr<SpawnStrategy> strategy);
+        void setUpStrategies(std::shared_ptr<GameplayInfoSource> gameplayInfoSource, std::shared_ptr<EnemyManager> enemyManager);
         void update(float deltaTime);
 };
